refactor(models): fix import structure for consistency
- Move models.py content into models/__init__.py
- Remove complex importlib machinery that could cause import issues
- Eliminates same architectural problem that was fixed in database/
- Maintains all existing functionality with cleaner import structure
- All tests pass and imports work correctly
